backend/requirements.txt
Flask==2.0.1
gunicorn==20.0.4
requests==2.25.1
psycopg2-binary==2.9.1
pandas==1.3.1
mysql-connector-python==8.4.0
packaging==24.0
pycparser==2.22
python-keycloak==4.1.0